+++ Magic/CHANGELOG.md
@@ -5,6 +5,7 @@
  - Add Magic Torch (magictorch) item, not craftable, requires LightAPI
  - Added support for spells requiring SkillAPI classes or skills. 
    Thanks @robotnikthingy and my other helpful contributors!
+ - Add commands and generic requirements support to Selector action
    
 ## 7.1.5